pyproject.toml
[project]
name = "pythonproject11"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "monsterui>=1.0.37",
    "python-fasthtml>=0.12.39",
]
